refractor: make componenents reusable and update Storybook stories
