+++ src/components/ui/molecule/CMSRating.jsx
@@ -1,18 +1,17 @@
 import StarRating from './starRating';
 
-export default function FacilityCMSRating() {
+export default function CMSRating({ stars = [] }) {
   return (
     <div>
       <div className="py-6">
         <div className="text-heading-xs">CMS Ratings</div>
       </div>
       <div className="grid grid-cols-1 sm:grid-cols-2">
-        <div>
-          <StarRating title="Overall Star Rating" rating={3.5} />
-        </div>
-        <div>
-          <StarRating title="Average Collective Owner Ranking" rating={4} />
-        </div>
+        {stars.map((stars, i) => (
+          <div key={stars.title + i}>
+            <StarRating title={stars.title} rating={stars.rating} />
+          </div>
+        ))}
       </div>
     </div>
   );
